MdeModulePkg PCD: Fix the issue to set the big SkuId
Contributed-under: TianoCore Contribution Agreement 1.1
Cc: Star Zeng <star.zeng@intel.com>
(cherry picked from commit f0c69b614cf56df1e7908574111d92864ca3ee9c)

+++ MdeModulePkg/Universal/PCD/Dxe/Service.c
@@ -788,7 +788,7 @@ UpdatePcdDatabase (
     SkuDelta = NULL;
     while (Index < mPeiPcdDbSize) {
       SkuDelta = (PCD_DATABASE_SKU_DELTA *) ((UINT8 *) mPeiPcdDbBinary + Index);
-      if (SkuDelta->SkuId == (UINT16) SkuId && SkuDelta->SkuIdCompared == 0) {
+      if (SkuDelta->SkuId == SkuId && SkuDelta->SkuIdCompared == 0) {
         break;
       }
       Index = (Index + SkuDelta->Length + 7) & (~7);
